Split orthohelp BDD step modules
Move command, cache, and IR assertions out of the shared `steps` module
so the behavioural test support is easier to navigate while preserving the
existing step text and scenario state fixture.

Treat a missing target directory as an empty orthohelp cache during cache
cleanup. Also update the bridge cache comment to use the project spelling
convention for normalization.
