+++ cargo-orthohelp/tests/rstest_bdd/behaviour/mod.rs
@@ -4,3 +4,6 @@ mod powershell_steps;
 mod roff_steps;
 mod scenarios;
 pub mod steps;
+mod steps_cache;
+mod steps_cmd;
+mod steps_ir;
