+++ cargo-orthohelp/tests/rstest_bdd/behaviour/steps_cache.rs
@@ -1,0 +1,155 @@
+//! Cache assertion helpers and step definitions for `cargo-orthohelp`.
+
+use std::io::Read;
+
+use camino::Utf8PathBuf;
+use cap_std::ambient_authority;
+use cap_std::fs_utf8::{Dir, DirEntry};
+use cap_std::time::SystemTime;
+use rstest_bdd_macros::then;
+
+use super::steps::{OrthoHelpContext, StepResult};
+use super::steps_cmd::get_workspace_root;
+
+pub(super) const CACHE_ARGS: &[&str] = &[
+    "--cache",
+    "--package",
+    "orthohelp_fixture",
+    "--locale",
+    "en-US",
+    "--locale",
+    "fr-FR",
+];
+
+#[then("the cached IR is reused")]
+fn cached_ir_reused(orthohelp_context: &mut OrthoHelpContext) -> StepResult<()> {
+    let cache_path = orthohelp_context
+        .cache_ir_path
+        .with_ref(Clone::clone)
+        .ok_or("cached IR path should be recorded")?;
+    let previous_content = orthohelp_context
+        .cache_ir_content
+        .with_ref(Clone::clone)
+        .ok_or("cached IR content should be recorded")?;
+    let cache_dir = cache_path.parent().ok_or("cached IR parent missing")?;
+    let file_name = cache_path.file_name().ok_or("cached IR filename missing")?;
+    let dir = Dir::open_ambient_dir(cache_dir, ambient_authority())?;
+    let mut file = dir.open(file_name)?;
+    let mut current_content = String::new();
+    file.read_to_string(&mut current_content)?;
+    assert_eq!(
+        previous_content, current_content,
+        "cached IR should not be rewritten"
+    );
+    Ok(())
+}
+
+#[then("the cached IR deserialises into the schema")]
+fn cached_ir_deserialises(orthohelp_context: &mut OrthoHelpContext) -> StepResult<()> {
+    let cache_path = orthohelp_context
+        .cache_ir_path
+        .with_ref(Clone::clone)
+        .ok_or("cached IR path should be recorded")?;
+    let cache_dir = cache_path.parent().ok_or("cached IR parent missing")?;
+    let file_name = cache_path.file_name().ok_or("cached IR filename missing")?;
+    let dir = Dir::open_ambient_dir(cache_dir, ambient_authority())?;
+    let mut file = dir.open(file_name)?;
+    let mut json = String::new();
+    file.read_to_string(&mut json)?;
+    let metadata: ortho_config::docs::DocMetadata = serde_json::from_str(&json)?;
+    assert_eq!(
+        metadata.ir_version,
+        ortho_config::docs::ORTHO_DOCS_IR_VERSION,
+        "cached IR should match the current schema version"
+    );
+    Ok(())
+}
+
+#[then("the command fails due to missing cache")]
+fn command_fails_due_to_missing_cache(orthohelp_context: &mut OrthoHelpContext) -> StepResult<()> {
+    let failed_with_missing_cache = orthohelp_context
+        .last_output
+        .with_ref(|output| {
+            let stderr = String::from_utf8_lossy(&output.stderr);
+            !output.status.success() && stderr.contains("MissingCache")
+        })
+        .ok_or("last_output should be set")?;
+    if !failed_with_missing_cache {
+        return Err("expected command failure due to missing cache".into());
+    }
+    Ok(())
+}
+
+pub(super) fn record_cache_state(ctx: &mut OrthoHelpContext) -> StepResult<()> {
+    let cache_path = find_cached_ir(ctx)?.ok_or("cached IR should exist")?;
+    let cache_dir = cache_path.parent().ok_or("cached IR parent missing")?;
+    let file_name = cache_path.file_name().ok_or("cached IR filename missing")?;
+    let dir = Dir::open_ambient_dir(cache_dir, ambient_authority())?;
+    // Read the full content: gives a stable reference for comparison and forces
+    // Windows NTFS to commit the final mtime before any subsequent metadata read.
+    let mut file = dir.open(file_name)?;
+    let mut content = String::new();
+    file.read_to_string(&mut content)?;
+    ctx.cache_ir_path.set(cache_path);
+    ctx.cache_ir_content.set(content);
+    Ok(())
+}
+
+fn find_cached_ir(ctx: &OrthoHelpContext) -> StepResult<Option<Utf8PathBuf>> {
+    let workspace_root = get_workspace_root(ctx)?;
+    // Respect CARGO_TARGET_DIR when set (e.g., by cargo-llvm-cov), because
+    // cargo-orthohelp inherits the variable and writes its cache relative to
+    // the effective target directory rather than the default `target/`.
+    let target_dir = std::env::var("CARGO_TARGET_DIR").map_or_else(
+        |_| workspace_root.join("target"),
+        |v| {
+            let p = Utf8PathBuf::from(&v);
+            if p.is_absolute() {
+                p
+            } else {
+                workspace_root.join(p)
+            }
+        },
+    );
+    let cache_root = target_dir.join("orthohelp");
+    let dir = match Dir::open_ambient_dir(&cache_root, ambient_authority()) {
+        Ok(d) => d,
+        Err(e) if is_not_found_kind(&e) => return Ok(None),
+        Err(e) => return Err(e.into()),
+    };
+    let mut newest: Option<(SystemTime, Utf8PathBuf)> = None;
+    for entry in dir.read_dir(".")? {
+        if let Some(candidate) = check_cache_entry(&cache_root, entry, newest.as_ref()) {
+            newest = Some(candidate);
+        }
+    }
+    Ok(newest.map(|(_, path)| path))
+}
+
+fn check_cache_entry(
+    cache_root: &Utf8PathBuf,
+    entry_result: Result<DirEntry, std::io::Error>,
+    newest: Option<&(SystemTime, Utf8PathBuf)>,
+) -> Option<(SystemTime, Utf8PathBuf)> {
+    let entry = entry_result.ok()?;
+    let file_type = entry.file_type().ok()?;
+    if !file_type.is_dir() {
+        return None;
+    }
+
+    let file_name = entry.file_name().ok()?;
+    let relative = Utf8PathBuf::from(file_name).join("ir.json");
+    let dir = Dir::open_ambient_dir(cache_root.as_path(), ambient_authority()).ok()?;
+    let metadata = dir.metadata(&relative).ok()?;
+    let modified = metadata.modified().ok()?;
+    let should_replace = newest.is_none_or(|(best_time, _)| modified > *best_time);
+    if !should_replace {
+        return None;
+    }
+
+    Some((modified, cache_root.join(relative)))
+}
+
+pub(super) fn is_not_found_kind(err: &std::io::Error) -> bool {
+    matches!(err.kind(), std::io::ErrorKind::NotFound)
+}
